move sample data tests to function for reuse
