+++ cypress/integration/opensearch-dashboards/with-security/login-page-spec.js
@@ -1,13 +1,16 @@
 import { LoginPage } from '@opensearch-dashboards-test/opensearch-dashboards-test-library'
+import { dashboardSanityTests } from '../../common/dashboard_sample_data_spec.js'
 
 const loginPage = new LoginPage(cy);
 
 describe('login into security enabled cluster', () => {
     beforeEach(() => {
-        cy.visit('/');
+        Cypress.Cookies.preserveOnce('security_authentication')
     })
 
     it('type user name and password to login', () => {
+        cy.visit('/');
+
         loginPage.enterUserName(Cypress.env('userName'))
 
         loginPage.enterPassword(Cypress.env('password'))
@@ -16,4 +19,6 @@ describe('login into security enabled cluster', () => {
 
         cy.url().should('include', '/app/home#/')
     })
+
+    dashboardSanityTests()
 })
